backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
